api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.5"
django = "^3.1"
djangorestframework = "^3.11.1"
psycopg2 = "^2.8.5"
django-environ = "^0.4.5"
markdown = "^3.2.2"
django-filter = "^2.3.0"
django-cacheops = "^5.0.1"
celery = "^4.4.7"
pytest-django = "^3.9.0"
djangorestframework-simplejwt = "^4.4.0"

[tool.poetry.dev-dependencies]
ipython = "^7.17.0"
black = "^19.10b0"
flake8 = "^3.8.3"
pre-commit = "^2.6.0"

[tool.black]
line-length = 79
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
